ondevice stop: sending the TERM signal once a second (instead of only once) to avoid some lingering thread to block process termination

+++ ondevice/commands/stop_cmd.py
@@ -32,14 +32,16 @@ def run():
         sys.stderr.write("Stopping ondevice daemon...")
         sys.stderr.flush()
 
-        try:
-            os.kill(pid, signal.SIGTERM)
-        except ProcessLookupError as e:
-            sys.stderr.write("\nondevice daemon isn't running (stale pid: {0})\n".format(pid))
-            return 2
-
         # effective timeout: 5sec (20*250ms)
         for i in range(20):
+            if i % 4 == 0:
+                # send a SIGTERM once a second
+                try:
+                    os.kill(pid, signal.SIGTERM)
+                except ProcessLookupError as e:
+                    sys.stderr.write("\nondevice daemon isn't running (stale pid: {0})\n".format(pid))
+                    return 2
+
             if not psutil.pid_exists(pid):
                 sys.stderr.write("done\n")
                 return 0 # stopped it successfully
